chore: disable spell checker auto-correction
The decision to initially enable the spell-checker's auto-correct feature was based on its successful implementation in prominent repositories such as uv, ruff, and pytest.

However, following concerns raised by @gazpachoking, a re-evaluation was undertaken.

In light of the risk of erroneous auto-corrections and considering the relatively minor effort required for manual fixes, we will be moving forward without this feature enabled. This change is to mitigate the risk of introducing errors into the codebase.

The spell-checker will continue to identify and flag potential typos, but the final decision to implement a correction will remain with the author to ensure accuracy.